backend/requirements.txt
absl-py==2.4.0
annotated-doc==0.0.4
anyio==4.12.1
astunparse==1.6.3
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.6
filelock==3.25.0
Flask==3.1.3
flask-cors==6.0.2
Flask-SQLAlchemy==3.1.1
flatbuffers==25.12.19
fsspec==2026.2.0
gast==0.7.0
google-pasta==0.2.0
greenlet==3.3.2
grpcio==1.78.0
h11==0.16.0
h5py==3.14.0
hf-xet==1.3.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.5.0
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
keras==3.13.2
libclang==18.1.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
namex==0.1.0
networkx==3.6.1
numpy==2.4.2
opt_einsum==3.4.0
optree==0.19.0
packaging==26.0
pluggy==1.6.0
protobuf==7.34.0
psycopg2-binary==2.9.11
Pygments==2.19.2
pytest==9.0.2
python-dotenv==1.2.2
PyYAML==6.0.3
regex==2026.2.28
requests==2.32.5
rich==14.3.3
safetensors==0.7.0
sentencepiece==0.2.1
setuptools==82.0.0
shellingham==1.5.4
six==1.17.0
SQLAlchemy==2.0.48
sympy==1.14.0
tensorflow==2.21.0
termcolor==3.3.0
tokenizers==0.22.2
torch==2.10.0
tqdm==4.67.3
transformers==5.2.0
typer==0.24.1
typer-slim==0.24.0
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.6
wheel==0.46.3
wrapt==2.1.2
